pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["walkquest/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
    "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.poetry]
name = "walkquest"
version = "0.1.0"
description = ""
authors = ["Andrea Maestri <16215183+andreamaestri@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
python-slugify = "8.0.4"
pillow = "11.0.0"
rcssmin = "1.1.2"
argon2-cffi = "23.1.0"
whitenoise = "6.8.2"
redis = "5.2.0"
hiredis = "3.0.0"
celery = "5.4.0"
django-celery-beat = "2.7.0"
django = ">=4.2.16"
django-environ = "0.11.2"
django-model-utils = "5.0.0"
django-allauth = {version = "^65.4.1", extras = ["mfa"]}
django-crispy-forms = "2.3"
crispy-bootstrap5 = "2024.10"
django-compressor = "4.5.1"
django-redis = "5.4.0"
gunicorn = {version = "23.0.0", extras = ["production"]}
psycopg = {version = "3.2.3", extras = ["production"]}
django-anymail = {version = "12.0", extras = ["production"]}
django-unfold = "^0.42.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"
django-tagulous = "^2.1.0"
rapidfuzz = "^3.10.1"
geopy = "^2.4.1"
django-extensions = "^3.2.3"
slippers = "^0.6.2"
django-esm = "^0.4.2"
django-storages = "^1.14.4"
boto3 = "^1.35.79"
django-ninja = "^1.3.0"
pydantic = "^2.10.3"
django-allauth-ui = "^1.5.3"
django-widget-tweaks = "^1.5.0"
django-browser-reload = "^1.17.0"
orjson = "^3.10.15"
django-vite = "^3.0.6"


[tool.poetry.group.dev.dependencies]
werkzeug = {version = "3.1.3", extras = ["watchdog"]}
ipdb = "0.13.13"
psycopg = {version = "3.2.3", extras = ["binary"]}
watchfiles = "0.24.0"
mypy = "1.13.0"
django-stubs = {version = "5.1.1", extras = ["compatible-mypy"]}
pytest = "8.3.3"
pytest-sugar = "1.0.0"
sphinx = "8.1.3"
sphinx-autobuild = "2024.10.3"
ruff = "0.8.0"
coverage = "7.6.5"
djlint = "1.36.1"
pre-commit = "4.0.1"
factory-boy = "3.3.1"
django-debug-toolbar = "4.4.6"
django-extensions = "3.2.3"
django-coverage-plugin = "3.1.0"
pytest-django = "4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
